design: 게시글 삭제 다이얼로그 추가

+++ feature/home/src/main/java/com/example/home/graph/post/component/PostDropDownMenu.kt
@@ -10,6 +10,10 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -17,6 +21,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.example.common.util.clickable
 import com.example.designsystem.R
+import com.example.designsystem.component.CheckCancleDialog
 import com.example.designsystem.theme.TraceTheme
 import com.example.designsystem.theme.White
 
@@ -27,46 +32,66 @@ fun OwnPostDropdownMenu(
     onEdit: () -> Unit,
     onDelete: () -> Unit,
 ) {
-    DropdownMenu(
-        expanded = expanded,
-        onDismissRequest = onDismiss,
-        shape = RoundedCornerShape(8.dp),
-        modifier = Modifier
-            .background(White),
-    ) {
-        Row(
-            modifier = Modifier.clickable(isRipple = true) {
-                onDismiss()
-                onEdit()
-            }.padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Image(painter = painterResource(R.drawable.edit_ic), contentDescription = "신고하기")
+    var showDeleteDialog by remember { mutableStateOf(false) }
 
-            Spacer(Modifier.width(12.dp))
-
-            Text(stringResource(R.string.edit), style = TraceTheme.typography.bodyMR)
-
-            Spacer(Modifier.width(70.dp))
-        }
-
-        Row(
-            modifier = Modifier.clickable(isRipple = true) {
-                onDismiss()
+    if (showDeleteDialog) {
+        CheckCancleDialog(
+            onCheck = {
                 onDelete()
-            }.padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Image(painter = painterResource(R.drawable.delete_ic), contentDescription = "삭제하기")
-
-            Spacer(Modifier.width(12.dp))
-
-            Text(stringResource(R.string.delete), style = TraceTheme.typography.bodyMR)
-
-            Spacer(Modifier.width(70.dp))
-        }
-
+                showDeleteDialog = false
+            },
+            onDismiss = { showDeleteDialog = false },
+            dialogText = "정말 삭제하시겠습니까?"
+        )
     }
+
+    if (expanded) {
+        DropdownMenu(
+            expanded = expanded,
+            onDismissRequest = onDismiss,
+            shape = RoundedCornerShape(8.dp),
+            modifier = Modifier
+                .background(White),
+        ) {
+            Row(
+                modifier = Modifier
+                    .clickable(isRipple = true) {
+                        onDismiss()
+                        onEdit()
+                    }
+                    .padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Image(painter = painterResource(R.drawable.edit_ic), contentDescription = "신고하기")
+
+                Spacer(Modifier.width(12.dp))
+
+                Text(stringResource(R.string.edit), style = TraceTheme.typography.bodyMR)
+
+                Spacer(Modifier.width(70.dp))
+            }
+
+            Row(
+                modifier = Modifier
+                    .clickable(isRipple = true) {
+                        onDismiss()
+                        showDeleteDialog = true
+                    }
+                    .padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Image(painter = painterResource(R.drawable.delete_ic), contentDescription = "삭제하기")
+
+                Spacer(Modifier.width(12.dp))
+
+                Text(stringResource(R.string.delete), style = TraceTheme.typography.bodyMR)
+
+                Spacer(Modifier.width(70.dp))
+            }
+
+        }
+    }
+
 }
 
 @Composable
@@ -75,27 +100,31 @@ fun OtherPostDropdownMenu(
     onDismiss: () -> Unit,
     onReport: () -> Unit,
 ) {
-    DropdownMenu(
-        expanded = expanded,
-        onDismissRequest = onDismiss,
-        shape = RoundedCornerShape(8.dp),
-        modifier = Modifier
-            .background(White),
-    ) {
-        Row(
-            modifier = Modifier.clickable(isRipple = true) {
-                onDismiss()
-                onReport()
-            }.padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
-            verticalAlignment = Alignment.CenterVertically,
+    if (expanded) {
+        DropdownMenu(
+            expanded = expanded,
+            onDismissRequest = onDismiss,
+            shape = RoundedCornerShape(8.dp),
+            modifier = Modifier
+                .background(White),
         ) {
-            Image(painter = painterResource(R.drawable.report_ic), contentDescription = "신고하기")
+            Row(
+                modifier = Modifier
+                    .clickable(isRipple = true) {
+                        onDismiss()
+                        onReport()
+                    }
+                    .padding(top = 15.dp, bottom = 15.dp, start = 15.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Image(painter = painterResource(R.drawable.report_ic), contentDescription = "신고하기")
 
-            Spacer(Modifier.width(12.dp))
+                Spacer(Modifier.width(12.dp))
 
-            Text(stringResource(R.string.report), style = TraceTheme.typography.bodyMR)
+                Text(stringResource(R.string.report), style = TraceTheme.typography.bodyMR)
 
-            Spacer(Modifier.width(70.dp))
+                Spacer(Modifier.width(70.dp))
+            }
         }
     }
 }
